refactor: ProfileTypeをentities/profile/modelに移行
- src/shared/entities/profile.tsをsrc/entities/profile/model/types.tsに移動
- entities/profile/index.tsを作成してPublic APIを定義
- 全てのインポートパスをentities/profileに更新
- shared/entities/index.tsでentities/profileを再エクスポート
