+++ src/entities/profile/index.ts
@@ -1,0 +1,1 @@
+export type { ProfileType } from "./model/types";
