+++ src/pages/profile/ui/ProfileHeader.tsx
@@ -1,7 +1,7 @@
 import { Avatar, Center, Flex, Space, Text } from "@mantine/core";
 import { IconLink, IconMapPinPin } from "@tabler/icons-react";
 
-import type { ProfileType } from "@/shared/entities";
+import type { ProfileType } from "@/entities/profile";
 
 interface ProfileHeaderProps {
 	user: ProfileType;
