+++ src/shared/entities/index.ts
@@ -1,5 +1,5 @@
 export type { User } from "./user";
 export type { PostType, PostProps } from "./post";
-export type { ProfileType } from "./profile";
+export type { ProfileType } from "../../entities/profile";
 export type { Follow } from "./follow";
 export type { Like } from "./like";
